corrections

+++ src/chmdocument.h
@@ -33,7 +33,7 @@ class CHMDocument : public Document
 public:
     CHMDocument();
     virtual int load(const QString &fileName);
-    virtual const QPixmap* getPixmap(int page, int scaleFactor);
+    virtual const QPixmap* getPixmap(int page, qreal scaleFactor);
     virtual void deletePixmap(const QPixmap *pixmap);
     virtual ~CHMDocument();
 private:
